+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - CHANNELS attribute to EXT-X-MEDIA
+- GetAllAlternatives() method to MasterPlaylist
 
 ### Fixed